# Make InexactError more informative