@timholy 2017-05-22
OK, I'm back to this. This time I think I'm really close, but the codegen part still isn't right: once it starts building inference.jl, memory spirals totally out of control and nearly brings my machine to a halt. Any tips would be most appreciated. You can ignore the julia part for now and just focus on `src/`, most likely `src/intrinsics.cpp`.

I should probably explain that, for ease of development, I decided to first create a new type, `InvalidValueError`, for the new error parametrization, and then use it in place of calls to `InexactError`. That let me have a buildable system at intermediate stages of development. (Assuming people prefer the original name, I'll search-replace InvalidValueError->InexactError at the end. Though perhaps there's some reason to consider a rename.)