@timholy 2017-07-09
Locally I'm seeing about 12 regressions compared to master. Only two of these seem real, and they're really just one because they are `x*y` and `y*x` of the following:
```julia
x = Complex{Int}(2)
y = Complex{UInt}(2)
```
The performance change is pretty serious, `@btime` shows 2.379 ns on master and 7.668 ns on this PR. Surprisingly, AFAICT it comes down to the registers assigned by LLVM.
Master:
```julia
julia> @code_native x*y
        .text
Filename: complex.jl
        pushq   %rbp
        movq    %rsp, %rbp
Source line: 404
        movq    (%rsi), %rax
        testq   %rax, %rax
        js      L65
Source line: 404
        movq    8(%rsi), %r8
        testq   %r8, %r8
        js      L87
Source line: 238
        movq    (%rdx), %rsi
        movq    8(%rdx), %rdx
        movq    %rsi, %rcx
        imulq   %rax, %rcx
        imulq   %rdx, %rax
        imulq   %r8, %rdx
        imulq   %r8, %rsi
        subq    %rdx, %rcx
        addq    %rax, %rsi
        movq    %rcx, (%rdi)
        movq    %rsi, 8(%rdi)
        movq    %rdi, %rax
        popq    %rbp
        retq
Source line: 404
L65:
        movabsq $jl_throw, %rax
        movabsq $jl_system_image_data, %rdi
        callq   *%rax
Source line: 404
L87:
        movabsq $jl_throw, %rax
        movabsq $jl_system_image_data, %rdi
        callq   *%rax
        nopl    (%rax)
```
This PR:
```julia
julia> @code_native x*y
        .text
Filename: complex.jl
        pushq   %rbp
        movq    %rsp, %rbp
        movq    %rdx, %rcx
Source line: 407
        movq    (%rsi), %rdx
        testq   %rdx, %rdx
        js      L68
Source line: 407
        movq    8(%rsi), %r8
        testq   %r8, %r8
        js      L102
Source line: 238
        movq    (%rcx), %rsi
        movq    8(%rcx), %rcx
        movq    %rsi, %rax
        imulq   %rdx, %rax
        imulq   %rcx, %rdx
        imulq   %r8, %rcx
        imulq   %r8, %rsi
        subq    %rcx, %rax
        addq    %rdx, %rsi
        movq    %rax, (%rdi)
        movq    %rsi, 8(%rdi)
        movq    %rdi, %rax
        popq    %rbp
        retq
Source line: 407
L68:
        movabsq $throw_inexacterror, %rax
        movabsq $140277663556160, %rdi  # imm = 0x7F94F04E6240
        movabsq $140277666121520, %rsi  # imm = 0x7F94F0758730
        callq   *%rax
        ud2
Source line: 407
L102:
        movabsq $throw_inexacterror, %rcx
        movabsq $140277663556160, %rdi  # imm = 0x7F94F04E6240
        movabsq $140277666121520, %rsi  # imm = 0x7F94F0758730
        movq    %r8, %rdx
        callq   *%rcx
        ud2
        nopl    (%rax,%rax)
```
Anything explicitly dealing with the new exception type is tucked away into the "goto", so it seems that in principle things should have been OK. I wonder if we could hint to LLVM that `throw_inexacterror` doesn't return?

But this is the only case I'm worried about. Of the rest:
- a bitshift change occurs because the test was being reduced to returning a constant, in the wake of #22210; that test should get an `@noinline` (I've checked that adding one makes the two perform similarly)
- in every other case, while we're worse than current master, we're nevertheless still equal to or better than 0.6. For example:
0.6:
```julia
julia> run(BaseBenchmarks.SUITE[["array", "index", ("sumrange", "BitArray{2}")]])
BenchmarkTools.Trial: 
  memory estimate:  109.38 KiB
  allocs estimate:  2000
  --------------
  minimum time:     138.231 μs (0.00% GC)
  median time:      146.327 μs (0.00% GC)
  mean time:        184.556 μs (20.24% GC)
  maximum time:     8.005 ms (97.58% GC)
  --------------
  samples:          5384
  evals/sample:     1
```
Master:
```julia
julia> run(BaseBenchmarks.SUITE[["array", "index", ("sumrange", "BitArray{2}")]])
BenchmarkTools.Trial: 
  memory estimate:  101.56 KiB
  allocs estimate:  1500
  --------------
  minimum time:     36.307 μs (0.00% GC)
  median time:      39.758 μs (0.00% GC)
  mean time:        75.239 μs (46.52% GC)
  maximum time:     10.656 ms (99.56% GC)
  --------------
  samples:          10000
  evals/sample:     1
```
This PR:
```julia
julia> run(BaseBenchmarks.SUITE[["array", "index", ("sumrange", "BitArray{2}")]])
BenchmarkTools.Trial: 
  memory estimate:  109.38 KiB
  allocs estimate:  2000
  --------------
  minimum time:     113.097 μs (0.00% GC)
  median time:      118.603 μs (0.00% GC)
  mean time:        159.162 μs (24.01% GC)
  maximum time:     14.161 ms (98.98% GC)
  --------------
  samples:          6189
  evals/sample:     1
```
So we give up a few of the gains from #22210, but we're still doing pretty well. (It just comes down to the `@noinline`d call to `throw_inexacterror` being penalized for 20 cycles of cost which prevents the BitArray constructor from being inlined. Again this could be solved if we knew that call never returned so we didn't have to penalize it.)

For these reasons, I think they only thing we have to consider is the case of `Complex{Int}*Complex{UInt}`, and whether that's something we need to worry about.